adds 'url' to log context for rendering errors (#687)
When Amphora throws an error in a location where req/res context is
available we don't currently include that context.

This commit includes 'url' in the log context to make it easy to
reproduce errors.

---

While `res.locals` is an object created by Express, it's not impossible
to manually delete or override that attribute. Using `_.get(res,'locals.url')`
will provide some additional safety to the logging
mechanism in the event of misuse.