reverts breaking upgrades from v7.10.0
A breaking upgrade to `handlebars` was released in error as v7.10.0.
This change was not intended for the public and v7.10.0 was removed from
the 'latest' tag on NPM / marked as deprecated.

This commit restores the `package.json` and `package-lock.json` files
that were present on the last valid release, v7.9.0.